some tests moved to no_version
